requirements.txt
fastapi
uvicorn
pymongo
pymongo[srv]
python-dotenv
pydantic
datetime
os
bson
pathlib
urllib